drop *gulp* support, server, global installation
As *geoffrey* is supposed to be run as CoffeeScript on the server
directly, there is no need for a build system converting the
CoffeeScript files into JavaScript.

*goeffrey* can be run as a server now. It listens for HTTP requests,
passing the questions through *wit.ai* and then call the right plugin.

Furthermore, a global installation via `nom install -g geoffrey` is now
supported.